backend-python/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
segmentation-models-pytorch
pillow
opencv-python-headless
numpy
python-multipart
albumentations